Restructured some exports
